update ghn and vnp

+++ src/main/java/com/example/backend/telosys/persistence/repositories/RatingsRepository.java
@@ -31,7 +31,7 @@ public interface RatingsRepository extends JpaRepository<Ratings, RatingsId> {
 
 	// Insert specific finders here
 
-	List<Ratings> findByBookId(int bookId);
+	List<Ratings> findByBookId(Long bookId);
 
 	// List<Ratings> findByXxxStartingWith(String xxx);
